feat: gradients
