+++ package.json
@@ -29,7 +29,11 @@
     "build:esm": "tsc -p tsconfig.esm.json",
     "build:cjs": "tsc -p tsconfig.cjs.json",
     "test": "mocha -r ts-node/register index.test.ts",
-    "print": "node -e \"const keys = require('./snapshots/index.test'); Object.keys(keys).map(key => console.log(key + '\\n' + keys[key].join('\\n'))) \" "
+    "print": "node -e \"const keys = require('./snapshots/index.test'); Object.keys(keys).map(key => console.log(key + '\\n' + keys[key].join('\\n'))) \" ",
+    "print:tests": "node -e \"const fs = require('fs'); const path = require('path'); const dir = './tests/snapshots'; fs.readdirSync(dir).filter(f => f.endsWith('.json')).forEach(file => { console.log('\\n=== ' + file + ' ===\\n'); const data = require(path.join(process.cwd(), dir, file)); Object.keys(data).forEach(key => console.log(key + '\\n' + data[key].join('\\n'))); });\"",
+    "print:test": "node -e \"const name = process.argv[1]; const data = require('./tests/snapshots/' + name + '.test.json'); Object.keys(data).forEach(key => console.log(key + '\\n' + data[key].join('\\n')));\" --"
+
+
   },
   "keywords": [
     "bar-chart",
