+++ tests/snapshots/horizontal-alignment.test.json
@@ -1,0 +1,78 @@
+{
+  "horizontal align top": [
+    "",
+    "A \u001b[0m╢██████████████████████",
+    "  \u001b[0m║",
+    "B \u001b[0m╢██████████████████████████████████████████████████",
+    "  \u001b[0m║",
+    "C \u001b[0m╢████████████████████████████████████",
+    "  \u001b[0m║",
+    "  \u001b[0m║",
+    "  \u001b[0m║",
+    "  \u001b[0m║",
+    "  \u001b[0m║",
+    "  \u001b[0m║",
+    "  \u001b[0m║",
+    "  \u001b[0m║",
+    "  \u001b[0m║",
+    "  \u001b[0m║",
+    "  ╚══════════════════════════════════════════════════"
+  ],
+  "horizontal align bottom": [
+    "",
+    "  \u001b[0m║",
+    "  \u001b[0m║",
+    "  \u001b[0m║",
+    "  \u001b[0m║",
+    "  \u001b[0m║",
+    "  \u001b[0m║",
+    "  \u001b[0m║",
+    "  \u001b[0m║",
+    "  \u001b[0m║",
+    "  \u001b[0m║",
+    "A \u001b[0m╢██████████████████████",
+    "  \u001b[0m║",
+    "B \u001b[0m╢██████████████████████████████████████████████████",
+    "  \u001b[0m║",
+    "C \u001b[0m╢████████████████████████████████████",
+    "  ╚══════════════════════════════════════════════════"
+  ],
+  "horizontal align center": [
+    "",
+    "  \u001b[0m║",
+    "  \u001b[0m║",
+    "  \u001b[0m║",
+    "  \u001b[0m║",
+    "  \u001b[0m║",
+    "A \u001b[0m╢██████████████████████",
+    "  \u001b[0m║",
+    "B \u001b[0m╢██████████████████████████████████████████████████",
+    "  \u001b[0m║",
+    "C \u001b[0m╢████████████████████████████████████",
+    "  \u001b[0m║",
+    "  \u001b[0m║",
+    "  \u001b[0m║",
+    "  \u001b[0m║",
+    "  \u001b[0m║",
+    "  ╚══════════════════════════════════════════════════"
+  ],
+  "horizontal align justify": [
+    "",
+    "A \u001b[0m╢██████████████████████",
+    "  \u001b[0m║",
+    "  \u001b[0m║",
+    "  \u001b[0m║",
+    "  \u001b[0m║",
+    "  \u001b[0m║",
+    "  \u001b[0m║",
+    "B \u001b[0m╢██████████████████████████████████████████████████",
+    "  \u001b[0m║",
+    "  \u001b[0m║",
+    "  \u001b[0m║",
+    "  \u001b[0m║",
+    "  \u001b[0m║",
+    "  \u001b[0m║",
+    "C \u001b[0m╢████████████████████████████████████",
+    "  ╚══════════════════════════════════════════════════"
+  ]
+}
