+++ tests/snapshots/vertical-alignment.test.json
@@ -1,0 +1,86 @@
+{
+  "vertical align justify default": [
+    "║                        █                                               ",
+    "║                        █                                               ",
+    "║                        █                       █                       ",
+    "║                        █                       █                       ",
+    "║                        █                       █                       ",
+    "║█                       █                       █                       ",
+    "║█                       █                       █                       ",
+    "║█                       █                       █                       ",
+    "║█                       █                       █                       ",
+    "║█                       █                       █                       ",
+    "╚═════════════════════════════════════════════════",
+    " A                       B                       C                       "
+  ],
+  "vertical align justify explicit": [
+    "║                        █                                               ",
+    "║                        █                                               ",
+    "║                        █                       █                       ",
+    "║                        █                       █                       ",
+    "║                        █                       █                       ",
+    "║█                       █                       █                       ",
+    "║█                       █                       █                       ",
+    "║█                       █                       █                       ",
+    "║█                       █                       █                       ",
+    "║█                       █                       █                       ",
+    "╚═════════════════════════════════════════════════",
+    " A                       B                       C                       "
+  ],
+  "vertical align left": [
+    "║   █                                                ",
+    "║   █                                                ",
+    "║   █  █                                             ",
+    "║   █  █                                             ",
+    "║   █  █                                             ",
+    "║█  █  █                                             ",
+    "║█  █  █                                             ",
+    "║█  █  █                                             ",
+    "║█  █  █                                             ",
+    "║█  █  █                                             ",
+    "╚══════════════════════════════════════════════════",
+    " A  B  C  "
+  ],
+  "vertical align right": [
+    "║                                              █     ",
+    "║                                              █     ",
+    "║                                              █  █  ",
+    "║                                              █  █  ",
+    "║                                              █  █  ",
+    "║                                           █  █  █  ",
+    "║                                           █  █  █  ",
+    "║                                           █  █  █  ",
+    "║                                           █  █  █  ",
+    "║                                           █  █  █  ",
+    "╚══════════════════════════════════════════════════",
+    "                                            A  B  C  "
+  ],
+  "vertical align center": [
+    "║                        █                           ",
+    "║                        █                           ",
+    "║                        █  █                        ",
+    "║                        █  █                        ",
+    "║                        █  █                        ",
+    "║                     █  █  █                        ",
+    "║                     █  █  █                        ",
+    "║                     █  █  █                        ",
+    "║                     █  █  █                        ",
+    "║                     █  █  █                        ",
+    "╚══════════════════════════════════════════════════",
+    "                      A  B  C  "
+  ],
+  "vertical align center stacked": [
+    "║                       \u001b[38;5;160m█  \u001b[0m\u001b[38;5;160m█  \u001b[0m                       ",
+    "║                       \u001b[38;5;160m█  \u001b[0m\u001b[38;5;160m█  \u001b[0m                       ",
+    "║                       \u001b[38;5;227m█  \u001b[0m\u001b[38;5;160m█  \u001b[0m                       ",
+    "║                       \u001b[38;5;227m█  \u001b[0m\u001b[38;5;227m█  \u001b[0m                       ",
+    "║                       \u001b[38;5;227m█  \u001b[0m\u001b[38;5;227m█  \u001b[0m                       ",
+    "║                       \u001b[38;5;227m█  \u001b[0m\u001b[38;5;227m█  \u001b[0m                       ",
+    "║                       \u001b[38;5;227m█  \u001b[0m\u001b[32m█  \u001b[0m                       ",
+    "║                       \u001b[32m█  \u001b[0m\u001b[32m█  \u001b[0m                       ",
+    "║                       \u001b[32m█  \u001b[0m\u001b[32m█  \u001b[0m                       ",
+    "║                       \u001b[32m█  \u001b[0m\u001b[32m█  \u001b[0m                       ",
+    "╚══════════════════════════════════════════════════",
+    "                        Q1 Q2 "
+  ]
+}
